gm: split all upstream logic into separate file
