+++ pkg/globalmanager/controllers/jointinference/jointinferenceservice.go
@@ -533,66 +533,6 @@ func (c *Controller) createEdgeWorker(service *sednav1.JointInferenceService, bi
 	return err
 }
 
-func (c *Controller) updateMetrics(name, namespace string, metrics []sednav1.Metric) error {
-	client := c.client.JointInferenceServices(namespace)
-
-	return runtime.RetryUpdateStatus(name, namespace, func() error {
-		joint, err := client.Get(context.TODO(), name, metav1.GetOptions{})
-		if err != nil {
-			return err
-		}
-		joint.Status.Metrics = metrics
-		_, err = client.UpdateStatus(context.TODO(), joint, metav1.UpdateOptions{})
-		return err
-	})
-}
-
-// updateFromEdge syncs the edge updates to k8s
-func (c *Controller) updateFromEdge(name, namespace, operation string, content []byte) error {
-	// Output defines owner output information
-	type Output struct {
-		ServiceInfo map[string]interface{} `json:"ownerInfo"`
-	}
-
-	var status struct {
-		// Phase always should be "inference"
-		Phase  string  `json:"phase"`
-		Status string  `json:"status"`
-		Output *Output `json:"output"`
-	}
-
-	err := json.Unmarshal(content, &status)
-	if err != nil {
-		return err
-	}
-
-	// TODO: propagate status.Status to k8s
-
-	output := status.Output
-	if output == nil || output.ServiceInfo == nil {
-		// no output info
-		klog.Warningf("empty status info for joint inference service %s/%s", namespace, name)
-		return nil
-	}
-
-	info := output.ServiceInfo
-
-	for _, ignoreTimeKey := range []string{
-		"startTime",
-		"updateTime",
-	} {
-		delete(info, ignoreTimeKey)
-	}
-
-	metrics := runtime.ConvertMapToMetrics(info)
-
-	err = c.updateMetrics(name, namespace, metrics)
-	if err != nil {
-		return fmt.Errorf("failed to update metrics, err:%+w", err)
-	}
-	return nil
-}
-
 // New creates a new JointInferenceService controller that keeps the relevant pods
 // in sync with their corresponding JointInferenceService objects.
 func New(cc *runtime.ControllerContext) (runtime.FeatureControllerI, error) {
@@ -640,7 +580,7 @@ func New(cc *runtime.ControllerContext) (runtime.FeatureControllerI, error) {
 	jc.podStore = podInformer.Lister()
 	jc.podStoreSynced = podInformer.Informer().HasSynced
 
-	cc.UpstreamController.Add(KindName, jc.updateFromEdge)
+	jc.addUpstreamHandler(cc)
 
 	return jc, nil
 }
